25/09/09 memberRoutes 통합테스트 중

+++ src/utils/fileUtils.js
@@ -5,10 +5,13 @@ const profilePath = process.env.PROFILE_FILE_PATH;
 const boardPath = process.env.BOARD_FILE_PATH;
 
 
-export const getResizeProfileName = (filename) => `${filename}_300`;
+export const getResizeProfileName = (filename) => `${filename}`;
 
 export const deleteImageFile = (filename, fileType) => {
 	const filePath = fileType === 'profile' ? profilePath : boardPath;
-	fs.unlinkSync(path.join(filePath, filename));
+	console.error('deleteImageFile log!!!!!!!!!!!!!!!!!!!!');
+
+	if(process.env.NODE_ENV !== 'test')
+		fs.unlinkSync(path.join(filePath, filename));
 }
